isolate all server handlers into a single struct, without using global variables
